wip: start moving traits to maybe-uring
