+++ crates/maybe-uring/src/buf/tokio_uring_compat.rs
@@ -1,13 +1,39 @@
-unsafe impl<T> tokio_uring::buf::IoBuf for T
+#[repr(transparent)]
+pub struct BufCompat<T>(T)
 where
-    T: crate::IoBuf,
+    T: crate::buf::IoBuf;
+
+unsafe impl<T> tokio_uring::buf::IoBuf for BufCompat<T>
+where
+    T: crate::buf::IoBuf,
 {
-    // TODO: forward methods
+    #[inline(always)]
+    fn stable_ptr(&self) -> *const u8 {
+        self.0.stable_ptr()
+    }
+
+    #[inline(always)]
+    fn bytes_init(&self) -> usize {
+        self.0.bytes_init()
+    }
+
+    #[inline(always)]
+    fn bytes_total(&self) -> usize {
+        self.0.bytes_total()
+    }
 }
 
-unsafe impl<T> tokio_uring::buf::IoBufMut for T
+unsafe impl<T> tokio_uring::buf::IoBufMut for BufCompat<T>
 where
-    T: crate::IoBufMut,
+    T: crate::buf::IoBufMut,
 {
-    // TODO: forward methods
+    #[inline(always)]
+    fn stable_mut_ptr(&mut self) -> *mut u8 {
+        self.0.stable_mut_ptr()
+    }
+
+    #[inline(always)]
+    unsafe fn set_init(&mut self, pos: usize) {
+        self.0.set_init(pos)
+    }
 }
